+++ src/renderer/App.tsx
@@ -6,7 +6,7 @@ import { ErrorPage } from "@/renderer/ErrorPage";
 import FlowChartTab from "./routes/flow_chart/FlowChartTabView";
 import TestSequencerTab from "./routes/test_sequencer_panel/TestSequencerView";
 import DeviceTab from "./routes/device_panel/DeviceView";
-import { useTheme } from "@/renderer/providers/themeProvider";
+import { useTheme } from "@/renderer/providers/ThemeProvider";
 import { Layout } from "./routes/common/Layout";
 import { Index } from "./routes/index";
 import packageJson from "../../package.json";
