+++ src/renderer/components/plotly/PlotlyComponent.tsx
@@ -4,7 +4,7 @@ import Plotly from "plotly.js";
 import { plotLayout } from "./layout";
 import { useMemo } from "react";
 import { PlotProps } from "@/renderer/types/plotly";
-import { useTheme } from "@/renderer/providers/themeProvider";
+import { useTheme } from "@/renderer/providers/ThemeProvider";
 import _ from "lodash";
 
 const MATRIX_SIZE = {
