+++ src/renderer/context/testSequencerWS.context.tsx
@@ -3,7 +3,7 @@ import { SendJsonMessage } from "react-use-websocket/dist/lib/types";
 import { useTestSequencerState } from "@/renderer/hooks/useTestSequencerState";
 import { useEffect } from "react";
 import useWebSocket, { ReadyState } from "react-use-websocket";
-import { BackendMsg, Test } from "@/renderer/types/testSequencer";
+import { BackendMsg, Test } from "@/renderer/types/test-sequencer";
 import { mapToTestResult } from "@/renderer/routes/test_sequencer_panel/utils/TestUtils";
 import { toast } from "sonner";
 import { env } from "@/env";
