+++ src/renderer/routes/common/DarkModeToggle.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/renderer/components/ui/button";
 import { Moon, Sun } from "lucide-react";
-import { useTheme } from "@/renderer/providers/themeProvider";
+import { useTheme } from "@/renderer/providers/ThemeProvider";
 
 export const DarkModeToggle = () => {
   const { theme, setTheme } = useTheme();
