+++ src/renderer/routes/test_sequencer_panel/components/dnd/DraggableRow.tsx
@@ -5,7 +5,7 @@ import {
   TestSequenceDropResult,
 } from "@/renderer/routes/test_sequencer_panel/models/drag_and_drop";
 import { TableCell, TableRow } from "@/renderer/components/ui/table";
-import { TestSequenceElement } from "@/renderer/types/testSequencer";
+import { TestSequenceElement } from "@/renderer/types/test-sequencer";
 import { Row, flexRender } from "@tanstack/react-table";
 import { parseInt } from "lodash";
 import { useTestSequencerState } from "@/renderer/hooks/useTestSequencerState";
