--- src/renderer/utils/RandomPositionOffset.ts
@@ -1,9 +1,0 @@
-export const addRandomPositionOffset = (
-  pos: { x: number; y: number },
-  range: number,
-) => {
-  return {
-    x: pos.x + (Math.random() - 0.5) * range,
-    y: pos.y + (Math.random() - 0.5) * range,
-  };
-};
