+++ src/renderer/utils/array.ts
