+++ src/renderer/utils/gallery.ts
