+++ src/renderer/utils/text-wrap.ts
